article-删除补充二次确认

+++ src/views/airticle.vue
@@ -235,17 +235,32 @@ export default {
     // 删除
     onDelete: function (id) {
       console.log(id)
-      this.$axios({
-        method: 'DELETE',
-        url: `/articles/${id}`
+      this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        this.$message({
+          type: 'success',
+          message: '删除成功!'
+        })
+        this.$axios({
+          method: 'DELETE',
+          url: `/articles/${id}`
+        })
+          .then(res => {
+            console.log(res)
+            this.loadAirticle(this.currentPage)
+          })
+          .catch(err => {
+            console.log(err)
+          })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
-        .then(res => {
-          console.log(res)
-          this.loadAirticle(this.currentPage)
-        })
-        .catch(err => {
-          console.log(err)
-        })
     }
   },
   created () {
